Change capitalization for Head of Data role
I changed all occurrences of "head of data" to "company's Head of Data" to avoid confusion. Given the data science context of the notebook, with "head of data" in lowercase I thought it meant the first few lines of a dataset, but it actually means a person at the company.